requirements.txt
streamlit
pandas
plotly.express
datetime
numpy